requirements.txt
numpy==1.21.4
streamlit==1.3.0
pandas==1.3.5
scikit-learn==1.0.1
scipy==1.7.3
matplotlib==3.5.1
tensorflow-cpu==2.7.0
keras==2.7.0